PMM-7781 Update pmm dependency, eliminate redundant type casting

+++ api-tests/management/ia/templates_test.go
@@ -96,7 +96,7 @@ func assertTemplate(t *testing.T, expectedTemplate alert.Template, listTemplates
 				value, err := expectedParam.GetValueForFloat()
 				require.NoError(t, err)
 				assert.True(t, param.Float.HasDefault)
-				assert.Equal(t, float32(value), param.Float.Default)
+				assert.Equal(t, value, param.Float.Default)
 			}
 
 			if len(expectedParam.Range) != 0 {
@@ -104,8 +104,8 @@ func assertTemplate(t *testing.T, expectedTemplate alert.Template, listTemplates
 				require.NoError(t, err)
 				assert.True(t, param.Float.HasMax)
 				assert.True(t, param.Float.HasMin)
-				assert.Equal(t, float32(min), param.Float.Min)
-				assert.Equal(t, float32(max), param.Float.Max)
+				assert.Equal(t, min, param.Float.Min)
+				assert.Equal(t, max, param.Float.Max)
 			}
 
 			assert.Nil(t, param.Bool)
